refactor: 검색 state를 커스텀으로 분리
